Completed GUI
This has a working python GUI running three threads: (1) a python
server, (2) a serial connection that inputs urine flow readings and
outputs the flow rate for the mannequin, (3) the GUI interface. There is
also code to save .csv containing data, autoconnect to the mbed / the
ability to send error messages when no mbed is connected and display
charts using the HighCharts javascript interface with continuously
updating data and other optimal settings. The HighCharts code interfaces
with the server via a .JSON file that is posted to the server, and the
script for HighCharts is similarly hosted on the local server.